+++ data_disk.tf
@@ -1,5 +1,5 @@
 resource "azurerm_managed_disk" "data_disk" {
-  for_each                   = var.data_disks
+  for_each                   = var.is_imported ? 0 : var.var.data_disks
   name                       = lookup(var.name_overrides.data_disks, each.key, "disk-${var.virtual_machine_config.hostname}-${each.key}")
   location                   = var.virtual_machine_config.location
   resource_group_name        = var.resource_group_name
@@ -14,6 +14,7 @@ resource "azurerm_managed_disk" "data_disk" {
   disk_iops_read_only        = each.value["disk_iops_read_only"]
   disk_mbps_read_only        = each.value["disk_mbps_read_only"]
   max_shares                 = each.value["max_shares"] 
+  trusted_launch_enabled     = each.value["trusted_launch_enabled"]
 
   tags = merge(
     var.tags,
@@ -21,6 +22,38 @@ resource "azurerm_managed_disk" "data_disk" {
   )
   lifecycle {
     prevent_destroy = true
+  }
+}
+
+resource "azurerm_managed_disk" "data_disk" {
+  for_each                   = var.is_imported ? var.var.data_disks : 0
+  name                       = lookup(var.name_overrides.data_disks, each.key, "disk-${var.virtual_machine_config.hostname}-${each.key}")
+  location                   = var.virtual_machine_config.location
+  resource_group_name        = var.resource_group_name
+  storage_account_type       = each.value["storage_account_type"]
+  create_option              = each.value["create_option"]
+  source_resource_id         = each.value["source_resource_id"]
+  disk_size_gb               = each.value["disk_size_gb"]
+  on_demand_bursting_enabled = each.value["on_demand_bursting_enabled"]
+  zone                       = var.virtual_machine_config.zone
+  disk_iops_read_write       = each.value["disk_iops_read_write"]
+  disk_mbps_read_write       = each.value["disk_mbps_read_write"]
+  disk_iops_read_only        = each.value["disk_iops_read_only"]
+  disk_mbps_read_only        = each.value["disk_mbps_read_only"]
+  max_shares                 = each.value["max_shares"]
+  trusted_launch_enabled     = each.value["trusted_launch_enabled"]
+
+  tags = merge(
+    var.tags,
+    coalesce(each.value.tags, {})
+  )
+  lifecycle {
+    prevent_destroy = true
+    ignore_changes = [
+      upload_size_bytes,
+      create_option,
+      source_resource_id 
+    ]
   }
 }