+++ variables.tf
@@ -79,6 +79,20 @@ variable "subnet" {
   }
 }
 
+variable "is_imported" {
+  type = bool
+  nullable = false
+  default = false
+  description = <<-DOC
+  ```
+    Optianally. Specify whether the VM is imported from an existing VM. Defaults to false. 
+    If the VM is imported (true) the resource will not try to change the following properties: 
+    VM: identity, source_image_reference, source_image_id, timezone, admin_username, computer_name, admin_password 
+    Data Disk: upload_size_bytes, create_option, source_resource_id
+  ```
+  DOC
+}
+
 variable "virtual_machine_config" {
   type = object({
     hostname                                               = string
@@ -111,6 +125,15 @@ variable "virtual_machine_config" {
       ultra_ssd_enabled   = optional(bool, false)
       hibernation_enabled = optional(bool, false)
     }), {})
+
+    identity = optional(object({
+      identity_type = optional(string, null)
+      identity_ids  = optional(list(string), null)
+    }))
+
+    boot_diagnostics = optional(object({
+      storage_account_uri = optional(string, null)
+    }), {})
   })
 
   validation {
@@ -134,6 +157,10 @@ variable "virtual_machine_config" {
   validation {
     condition     = var.virtual_machine_config.zone != null ? var.virtual_machine_config.availability_set_id == null : true
     error_message = "Either 'zone' or 'availability_set_id' can be set, but not both."
+  }
+  validation {
+    condition     = var.virtual_machine_config.identity == null ? true : contains(["SystemAssigned", "UserAssigned", "SystemAssigned, UserAssigned"], var.virtual_machine_config.identity.identity_type)
+    error_message = "identity_type must be one of 'SystemAssigned', 'UserAssigned' or 'SystemAssigned, UserAssigned'."
   }
   description = <<-DOC
   ```
@@ -166,6 +193,11 @@ variable "virtual_machine_config" {
     additional_capabilities: (Optional) Additional capabilities for the virtual machine.
       ultra_ssd_enabled: (Optional) Enable UltraSSD_LRS for the virtual machine. Defaults to false.
       hibernation_enabled: (Optional) Enable hibernation for the virtual machine. Defaults to false.       
+    identity: (Optional) Identity configuration for the virtual machine.
+      identity_type: (Optional) The type of identity used for the Virtual Machine. Possible values are 'SystemAssigned', 'UserAssigned', 'SystemAssigned, UserAssigned' and null. Defaults to null.
+      identity_ids: (Optional) List of User Assigned Identity IDs when identity_type is set to 'UserAssigned' or 'SystemAssigned, UserAssigned'. Defaults to null.
+    boot_diagnostics: (Optional) Boot diagnostics configuration for the virtual machine.
+      storage_account_uri: (Optional) The URI of the storage account to use for boot diagnostics. Defaults to null.
   ```
   DOC
 }
@@ -203,6 +235,7 @@ variable "data_disks" {
     disk_mbps_read_only        = optional(number)
     max_shares                 = optional(number)
     tags                       = optional(map(string), {})
+    trusted_launch_enabled     = optional(bool, true)
   }))
   validation {
     condition     = length([for v in var.data_disks : v.lun]) == length(distinct([for v in var.data_disks : v.lun]))
@@ -394,12 +427,14 @@ variable "resource_group_name" {
 
 variable "name_overrides" {
   type = object({
-    nic             = optional(string)
-    nic_ip_config   = optional(string)
-    public_ip       = optional(string)
-    virtual_machine = optional(string)
-    os_disk         = optional(string)
-    data_disks      = optional(map(string), {})
+    nic                 = optional(string)
+    nic_ip_config       = optional(string)
+    public_ip           = optional(string)
+    virtual_machine     = optional(string)
+    os_disk             = optional(string)
+    data_disks          = optional(map(string), {})
+    hostname            = optional(string)
+    resource_group_name = optional(string)
   })
   description = "Possibility to override names that will be generated according to q.beyond naming convention."
   default     = {}